requirements.txt
satori-rtm-sdk
kafka-python
flask
